add translate to educator nav
