--- controllers/login.js
@@ -1,12 +1,0 @@
-const crypto = require('crypto');
-const  Course  = require('../models/classes/Course.js');// course not Course, Strange
-const  Student  = require('../models/classes/Student.js');
-const  Educator  = require('../models/classes/Educator.js');
-
-const Swal = require('sweetalert2');
-// console.log('in pages.js');
-const { usersCollection, coursesCollection } = require('../models/config');
-
-module.exports = {
-    
-}
