+++ index.js
@@ -28,7 +28,8 @@ app.set("view engine","ejs")
 // Add navbar Translator
 app.set("navsTranslator",{
   "Home":"/",
-  "Dashboard":"/admin",
+  "Dashboard Admin":"/admin",
+  "Dashboard Educator":"/EducatorDashboardMain",
   "Search":"/search",
   "Pricing":"/pricing",
   "Insert Course":"/EducatorDashboard"
@@ -98,9 +99,9 @@ app.use(express.urlencoded({extended:false}))
 // Rate Limit
 var limiter  = rateLimit({
   windowMs: 10*60*1000, // 1 hour window
-  delayAfter: 10, // begin slowing down responses after the first 10 requests
+  delayAfter: 20, // begin slowing down responses after the first 10 requests
   delayMs: 100, // slow down subsequent responses by 100 milliseconds per request
-  max: 50, // start blocking after 50 requests
+  max: 100, // start blocking after 50 requests
   standardHeaders: true, // Return rate limit info in the `RateLimit-*` headers
 	legacyHeaders: false, // Disable the `X-RateLimit-*` headers
   message: "Too many requests made from this IP, please try again in a few minutes"
