+++ models/classes/Admin.js
@@ -1,7 +1,7 @@
 const User = require('./User');
 class Admin extends User{
     constructor(name,email){
-        super(name,email,true,'admin','avatar12.svg',["Home","Dashboard"])
+        super(name,email,true,'admin','avatar12.svg',["Home","Dashboard Admin"])
     }
     addUser(){
         // some function
