+++ public/css/profile.css
@@ -1,3 +1,7 @@
+html[lang="ar"] div.textContent {
+  direction: rtl;
+}
+
 section .container {
   width: fit-content;
   padding: 30px;
@@ -10,6 +14,7 @@ section .container {
   background: rgba(50, 50, 50, 0.9);
   backdrop-filter: blur(20px);
 }
+
 .imgBox {
   cursor: pointer;
   position: relative;
