+++ routes/routes.js
@@ -2,6 +2,7 @@ const express= require('express')
 const router = express.Router();
 const controller = require('../controllers/pages')
 const controllerForget = require('../controllers/forget')
+const controllerAuth = require('../controllers/Auth')
 const controllerAdmin = require('../controllers/adminControllers')
 const rateLimit = require('express-rate-limit');
 console.log('in routes.js');
@@ -56,15 +57,15 @@ router.use('/admin',(req,res,next)=>{
     next()
 })
 router.get('/',controller.index)
-router.get('/login',checkUserLogin,controller.loginGet)
-router.post('/login',checkUserLogin,controller.loginPost)
-router.get('/register',checkUserLogin,controller.registerGet)
-router.post('/register',checkUserLogin,controller.registerPost)
+router.get('/login',checkUserLogin,controllerAuth.loginGet)
+router.post('/login',checkUserLogin,controllerAuth.loginPost)
+router.get('/register',checkUserLogin,controllerAuth.registerGet)
+router.post('/register',checkUserLogin,controllerAuth.registerPost)
+router.get('/signout',controllerAuth.signout)
 router.get('/forget',controllerForget.forgetGet)
 router.post('/forget',controllerForget.forgetPost)
 router.get('/reset-password/:token',controllerForget.resetPasswordGet)
 router.post('/reset-password/:token',controllerForget.resetPasswordPost)
-router.get('/signout',controller.signout)
 router.get('/search',controller.search)
 router.get('/profile',controller.profile)
 router.get('/pricing',controller.pricing)
