+++ views/EducatorDashboard.ejs
@@ -58,36 +58,7 @@
                       }) 
                 </script>
             <% } } %>
-        <nav class="navbar navbar-expand-lg navbar-dark p-lg-4 p-md-4"   style="background-color: #232323;">
-            <div class="container">
-              <a class="navbar-brand" href="/"><img class="logo" src="./logo.svg" alt="logo"></a>
-              <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
-                <span class="navbar-toggler-icon"></span>
-              </button>
-              <div class="collapse navbar-collapse" id="navbarSupportedContent"  style="background-color: #232323;">
-                <ul class="navbar-nav mb-2 mb-lg-0 w-100  align-items-center">
-                    <!-- <li class="nav-item"><a class="nav-link" href="/">Home</a></li> -->
-                    <li class="nav-item active"><a class="nav-link" aria-current="page" href="/">EducatorDashboard</a></li>
-                        <!-- <li class="nav-item loginLink"><a class="nav-link" href="/login">Login</a></li>
-                        <li class="nav-item"><a class="nav-link" href="/register">Register</a></li> -->
-                        <div class="dropstart ms-md-auto">
-                            <button class="btn dropdown-toggle d-flex justify-content-center align-items-center h-100" type="button" data-bs-toggle="dropdown" aria-expanded="false">
-                              <img src="/avatars/<%= data.user.img %>" class="rounded-circle navbar-user-avatar " style="width: 50px;">
-                            </button>
-
-                            <div class="dropdown-menu dropdown-menu-right" aria-labelledby="navbarAvatarDropdown">
-                            <a class="dropdown-item disabled" href="/">Hi <%= data.user.name %></a>
-                            <a class="dropdown-item" href="/profile">Profile</a>
-                            <a class="dropdown-item" href="https://twitter.com/intent/tweet?url=http://STEMREF.askardesign.com/&text=you+search+for+whatever+learning+matirial+here&hashtags=STEMREF" target="_blank">Refer to friends</a>
-                            <a class="dropdown-item" href="mailto:stemref@gmail.com">Give Feedback</a>
-                            <a class="dropdown-item" href="/signout">Signout <i class="fa-solid fa-arrow-right-from-bracket"></i></a>
-                        </div>
-                  
-                </ul>
-            
-              </div>
-            </div>
-          </nav>
+            <%- include('partials/header.ejs') %>
         <div class="box">
             <div class="container">
                 <form action="/EducatorDashboard" method="post">
