+++ views/courseContent.ejs
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html>
+<html lang="<%= req.language %>">
   <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
@@ -22,10 +22,8 @@
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Rubik:wght@300;500;700;900&family=Tulpen+One&display=swap" rel="stylesheet">
     <!-- Bootstrap CSS -->
-    <!-- <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-T3c6CoIi6uLrA9TneNEoa7RxnatzjcDSCmG1MXxSR1GAsXEV/Dwwykc2MPK8M2HN" crossorigin="anonymous"> -->
     <link rel="stylesheet" href="/libs/bootstrap.min.css"> 
     <!-- font awesome Library -->
-    <!-- <link href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.4.2/css/all.min.css" rel="stylesheet"> -->
     <link rel="stylesheet" href="/libs/all.css"> 
     <link rel="stylesheet" href="/libs/all.min.css"> 
     <!-- SweetAlert2 CSS -->
@@ -35,7 +33,7 @@
     <!-- CSS  -->
     <link rel="stylesheet" href="/css/style.css">
     <link rel="stylesheet" href="/css/courseContent.css">
-        <title>STEMREF | Course Content Page</title>
+      <title>STEMREF | Course Content Page</title>
   </head>
   <body>
     <main>
@@ -56,27 +54,18 @@
               <div class="descritpion" >
                 <h2>Description</h2>
                 <p><%= data.course.description %></p>
-                <!-- <h2>Tags:</h2>
-                <div class="description-tags">
-                  <% for(let tag of data.course.tags){ %>
-                    <div class="tag"><%= tag %></div>
-                  <% } %>
-                </div> -->
-
-
               </div>
               <% if((!data.catograySearch)){ %>
-                <a class="fa-solid fa-arrow-left" style="color: white;" href="/search"></a>
+                <a style="color: white;" href="/search"><i class="fa-solid fa-arrow-left"></i></a>
                 <% } %>
               <% if((data.catograySearch)){ %>
-                <a class="fa-solid fa-arrow-left" style="color: white;" href="/courseContent/<%=data.course.name%>"></a>
+                <a style="color: white;" href="/courseContent/<%=data.course.name%>"><i class="fa-solid fa-arrow-left"></i></a>
                 <% } %>
                 
           <div class="top-text">
             <h1><%= data.course.name %><span><%= data.catograySearch %></span></h1>
-
             <div class="tags">
-                <div class="tag">course</div>
+                <!-- <div class="tag">course</div> -->
                 <!-- <div class="tag">CS205</div> -->
                 <% for(let tag of data.course.tags){ %>
                   <div class="tag"><%= tag %></div>
@@ -84,9 +73,9 @@
             </div>
             <div class="rating">
               <% if(data.avg>0){ %>
-                <i class="fa-solid fa-star active "></i>
-                <%for(let i=1;i<data.avg;i++){%>
-                  <i class="fa-solid fa-star active"></i>
+                <i class="fa-solid fa-star <%= data.color %>"></i>
+                <%for(let i=1;i<Math.floor(data.avg);i++){%>
+                  <i class="fa-solid fa-star <%= data.color %>"></i>
                 <% } %>
                 <%for(let i=data.avg;i<5;i++){%>
                   <i class="fa-solid fa-star"></i>
@@ -132,7 +121,7 @@
               <div class="tools">
                 <button class="review-btn" title="Review"><i class="fa-solid fa-star btn"></i></button>
                 <!-- TODO: change review to popup review with message -->
-                <div class="stars hide">
+                <div class="stars hide ">
                   <input type="radio" name="start" id="star1" data-num-star="5"><label for="star1"><i class="fa-solid fa-star"></i></label>
                   <input type="radio" name="start" id="star2" data-num-star="4"><label for="star2"><i class="fa-solid fa-star"></i></label>
                   <input type="radio" name="start" id="star3" data-num-star="3"><label for="star3"><i class="fa-solid fa-star"></i></label>
