+++ views/partials/header.ejs
@@ -37,7 +37,7 @@
                     
                 <% } %>
             </ul>
-            <form action="#" method="get">
+            <form class="lang-form-select" action="#" method="get">
                 <select class="lang-select form-select " aria-label="select lang"  data-bs-theme="dark" name="lng" data-width="fit" onchange="this.form.submit()">
                     <option value="en" <% if(req.language=="en"){%> selected <%}%> >
                     EN
