+++ views/pricing.ejs
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html>
+<html lang="<%= req.language %>">
 
 <head>
     <meta charset="UTF-8">
