+++ views/profile.ejs
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html>
+<html lang="<%= req.language %>">
   <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
@@ -45,16 +45,16 @@
               </div>
                 <!-- <input class="disable" value="<%= data.user.name %>"> -->
                 <div class="textContent">
-                  <h3 class="disable" >Username: @<input type="text" value="<%= data.user.name %>"></h3>
-                <h3 class="disable" >Email: <input type="email" value="<%= data.user.email  %>"></h3>
-                <h3 class="disable" >Subscribe: <span></span><%= data.user.paidUser==true?data.user.paidUser:false %></span></h3>
+                  <h3 class="disable" ><%= t('LOGINPAGE.Username') %>: <input type="text" value="<%= data.user.name %>" placeholder="@"></h3>
+                <h3 class="disable" ><%= t('LOGINPAGE.Email') %>: <input type="email" value="<%= data.user.email  %>"></h3>
+                <h3 class="disable" ><%= t('PROFILEPAGE.Subscribe') %>: <span></span><%= t(`PROFILEPAGE.${data.user.paidUser==true?data.user.paidUser:false}`) %></span></h3>
                 <% if(data.user.reviewed){ %>
-                  <h3>Your Course Reviews:</h3>
+                  <h3><%= t('PROFILEPAGE.Your Reviews') %>:</h3>
                   <div class="reviewes">
                     <table class="table table-striped table-dark">
                       <thead>
-                        <th>Subject</th>
-                        <th>Reivew</th>
+                        <th><%= t('PROFILEPAGE.Subject') %></th>
+                        <th><%= t('PROFILEPAGE.Reivew') %></th>
                       </thead>
                       <tbody class="review">
                     <% for(let review of Object.keys(data.user.reviewed)){ %>
